maps media table to gsheets obs csv
I think media_data is returned as a DataFrame, so the parsing was very similar.

Left todo/cookies in place for more complete observation file if needed. This is just to fill in the template - Similar paths could be taken for media to observations fields - when we are farther along than just gsheets.